Nest the ul tag within an li tag
The existing file in the repository places a `<ul>` tag directly within a `<ul>` tag, which is invalid html.  This places the `<ul>` tag within an acceptable `<li>` tag.  This along with the associated CSS allows the display of the list to continue to be offset for items that are children of a previous title.  I have not discovered if/what additional scripting may be affected by this change nor do I believe that I have a ready way to test for that as my attempt to place a rewritten version of the html within a browser page did not pull up the associated CSS and therefore prevented observing operation.