Modified to bring in line with 1.5.7 (#295)
* Modified to bring in line with 1.5.7

I have made some changes to this page to add some more examples and make it so that i think it works. I have also added a bit on troubleshooting as i have been trying to make it work with the minimum of details. I was wondering if on installing a language pack we should first get the user to copy over all of the english files this way if the language pack is missing files or zen cart base has added new files they will not be missing when the language pack is installed. I do not know if this is the final version so please be as harsh as you like and let me know what needs changing so that i can get it right.

* Update languages.md

Co-authored-by: Scott C Wilson <scottcwilson@gmail.com>