Clarify that the Application secret is 'Production'
As the solution for 'The provided OAuth access token has expired' error message is also targeted at "non-technical store owners" it would be better to ensure the user is working on the Production tab of the Square app.

I believe that the first time the Squareup app page is accessed the tab defaults to Sandbox, for any subsequent visits the Squareup website remembers the users last tab selection - so they nmay or may not have the Production tab already selected.