Update changing_date_format.md
Have found a way to override the admin settings using extra_definitions

As these are called last they can override and of the admin values already defined.

I have changed it for date to UK format. I don't know if it would be worth adding something general for overriding admin settings. This will certainly save me a lot of checking work if admin settings are changed.