Attempt new file load if support new system
Without this change, the presence of the new file in absence of system support will generate an error. 

A similar change to 1.5.8+ versions would be beneficial to users but only in cases where new files were pushed against an older unsupported version. Therefore I don't see the value and am okay with the discrepancy between the differences.