Add minor history to snippet
The modification to this area maintains the at "worst" concept of working with the previous version; however, offers at least a little direct information about the reason for the code to be written the way it is. Unfortunately, it looks like it would take too many words to explain that placing both versions of the language file on the server causes the new version to be loaded rather than there being some sort of version dependency and loading the old file instead.

Would be better to prevent loading the new style until the original intended language/pre-language style didn't exist.