Update accessibility.md
The goal is to put information about concerns and tools but not make the page 40 pages long.

It's my intent to have separate pages for each bullet under Concerns and Tools.  That will allow each concern and/or tool to be kept current/correct without having to edit this page every time something changes.

While I can find the method for making the Concerns and Tools bullets, I am not familiar with the link to be used.  I will include a Lighthouse page with this PR as an example of what it would look like.