Homepage Revamp
I'm no designer but because it's something I enjoy doing I have revamped the Docs home page (and colour coordinated the `dev` & `user` headers). The only thing left to do would be higher resolution background image of the ZC logo, with added opacity (I can't achieve that with css).

Anyway, just throwing it out there in the hope that it adds value.